Conflicts:
	app/controllers/DocumentApiController.php

+++ app/controllers/DocController.php
@@ -29,8 +29,8 @@ class DocController extends BaseController{
 		try{
 			
 			//Retrieve requested document
-			$doc = Doc::where('slug', $slug)->with('statuses')->with('sponsor')->with('categories')->with('dates')->first();
-			
+			$doc = Doc::where('slug', $slug)->with('statuses')->with('userSponsor')->with('groupSponsor')->with('categories')->with('dates')->first();
+
 			if(!isset($doc)){
 				App::abort('404');
 			}
@@ -63,7 +63,7 @@ class DocController extends BaseController{
 			return View::make('doc.reader.index', $data);
 						
 		}catch(Exception $e){
-			return Redirect::to('docs')->with('error', $e->getMessage());
+			return Redirect::to('/')->with('error', $e->getMessage());
 		}
 		App::abort('404');
 	}
@@ -97,28 +97,27 @@ class DocController extends BaseController{
 
 		$supported = (bool)$input['support'];
 
-		$docMeta = DocMeta::where('user_id', Auth::user()->id)->where('meta_key', '=', 'support')->where('doc_id', '=', $doc)->first();
-
+		$docMeta = DocMeta::withTrashed()->where('user_id', Auth::user()->id)->where('meta_key', '=', 'support')->where('doc_id', '=', $doc)->first();
 
 		if(!isset($docMeta)){
 			$docMeta = new DocMeta();
-
 			$docMeta->doc_id = $doc;
 			$docMeta->user_id = Auth::user()->id;
 			$docMeta->meta_key = 'support';
 			$docMeta->meta_value = (string)$supported;
-
 			$docMeta->save();
 		}
-		elseif($docMeta->meta_value == (string)$supported){
+		elseif($docMeta->meta_value == (string)$supported && !$docMeta->trashed()){
 			$docMeta->delete();
 			$supported = null;
 		}else{
+			if ($docMeta->trashed()) {
+				$docMeta->restore();
+			}
 			$docMeta->doc_id = $doc;
 			$docMeta->user_id = Auth::user()->id;
 			$docMeta->meta_key = 'support';
-			$docMeta->meta_value = (string)$supported;
-
+			$docMeta->meta_value = (string)(bool)$input['support'];
 			$docMeta->save();
 		}
